Move CZT and FFS composition functions out of module, and into test scripts.
