+++ pyffs/czt.py
@@ -188,42 +188,3 @@ def cztn(Phi, A, W, M, axes=None):
 
     Phi_czt = g[time_idx]
     return Phi_czt
-
-
-def cztn_comp(Phi, A, W, M, axes=None):
-    """
-    Perform multi-dimensional CZT from signal samples of a D-dimension signal by performing D 1D
-    CZTs along each dimension.
-
-    Parameters
-    ----------
-    Phi : :py:class:`~numpy.ndarray`
-        (..., N_1, N_2, ..., N_D, ...) input values.
-    A : list(float or complex)
-        Circular offset from the positive real-axis, for each dimension.
-    W : list(float or complex)
-        Circular spacing between transform points, for each dimension.
-    M : list(int)
-        Length of transform for each dimension.
-    axes : tuple
-        Dimensions of `Phi` along which transform should be applied.
-
-    Returns
-    -------
-    X : :py:class:`~numpy.ndarray`
-        (..., M_1, M_2, ..., M_D, ...) transformed input along the axis indicated by `axes`.
-
-    Notes
-    -----
-    Due to numerical instability when using large `M`, this implementation only supports transforms
-    where `A` and `W` have unit norm.
-    """
-
-    axes, A, W = _verify_cztn_input(Phi, A, W, M, axes)
-
-    # sequence of 1D FFS
-    Phi_czt = Phi.copy()
-    for d, ax in enumerate(axes):
-        Phi_czt = czt(Phi_czt, A[d], W[d], M[d], axis=ax)
-
-    return Phi_czt
