+++ pyffs/ffs.py
@@ -134,155 +134,6 @@ def iffs(x_FS, T, T_c, N_FS, axis=-1):
     return iffsn(Phi_FS=x_FS, T=[T], T_c=[T_c], N_FS=[N_FS], axes=(axis,))
 
 
-def ffsn_comp(Phi, T, T_c, N_FS, axes=None):
-    r"""
-    Fourier Series coefficients from signal samples of a D-dimension signal by performing D 1D FFTs
-    along each dimension.
-
-    Parameters
-    ----------
-    Phi : :py:class:`~numpy.ndarray`
-        (..., N_s1, N_s2, ..., N_sD, ...) function values at sampling points specified by
-        :py:func:`~pyffs.util.ffsn_sample`.
-    T : list(float)
-        Function period along each dimension.
-    T_c : list(float)
-        Function bandwidth along each dimension.
-    N_FS : list(int)
-        Period mid-point for each dimension.
-    axes : tuple
-        Dimensions of `Phi` along which function samples are stored.
-
-    Returns
-    -------
-    Phi_FS : :py:class:`~numpy.ndarray`
-        (..., N_s1, N_s2, ..., N_sD, ...) array containing Fourier Series coefficients in ascending
-        order (top-left of matrix).
-
-    Examples
-    --------
-    Let :math:`\phi(x, y)` be a shifted Dirichlet kernel of periods :math:`(T_x, T_y)` and
-    bandwidths :math:`N_{FS, x} = 2 N_x + 1, N_{FS, y} = 2 N_y + 1`:
-
-    .. math::
-
-       \phi(x, y) &= \sum_{k_x = -N_x}^{N_x} \sum_{k_y = -N_y}^{N_y}
-                \exp\left( j \frac{2 \pi}{T_x} k_x (x - T_{c,x}) \right)
-                \exp\left( j \frac{2 \pi}{T_y} k_y (y - T_{c,y}) \right) \\
-               &= \frac{\sin\left( N_{FS, x} \pi [x - T_{c,x}] / T_x \right)}{\sin\left( \pi
-               [x - T_{c, x}] / T_x \right)} \frac{\sin\left( N_{FS, y} \pi [y - T_{c,y}] / T_y
-               \right)}{\sin\left( \pi [y - T_{c, y}] / T_y \right)}.
-
-    Its Fourier Series (FS) coefficients :math:`\phi_{k_x, k_y}^{FS}` can be analytically evaluated
-    using the shift-modulation theorem:
-
-    .. math::
-
-       \phi_{k_x, k_y}^{FS} =
-       \begin{cases}
-           \exp\left( -j \frac{2 \pi}{T_x} k_x T_{c,x} \right) \exp\left( -j \frac{2 \pi}{T_y} k_y
-           T_{c,y} \right) & -N_x \le k_x \le N_x, -N_y \le k_y \le N_y,  \\
-           0 & \text{otherwise}.
-       \end{cases}
-
-    Being bandlimited, we can use :py:func:`~pyffs.ffs.ffsn_comp` to numerically evaluate
-    :math:`\{\phi_{k_x, k_y}^{FS}, k_x = -N_x, \ldots, N_x, k_y = -N_y, \ldots, N_y\}`:
-
-    .. testsetup::
-
-       import math
-
-       import numpy as np
-
-       from pyffs import ffsn_sample, ffsn_comp
-       from pyffs.func import dirichlet_2D, dirichlet_fs
-
-    .. doctest::
-
-       >>> T = [1, 1]
-       >>> T_c = [0, 0]
-       >>> N_FS = [3, 3]
-       >>> N_s = [4, 3]
-
-       # Sample the kernel and do the transform.
-       >>> sample_points, _ = ffsn_sample(T=T, N_FS=N_FS, T_c=T_c, N_s=N_s)
-       >>> diric_samples = dirichlet_2D(sample_points, T, T_c, N_FS)
-       >>> diric_FS = ffsn_comp(Phi=diric_samples, T=T, N_FS=N_FS, T_c=T_c)
-
-       # Compare with theoretical result.
-       >>> diric_FS_exact = np.outer(
-       ... dirichlet_fs(N_FS[0], T[0], T_c[0]), dirichlet_fs(N_FS[1], T[1], T_c[1])
-       ... )
-       >>> np.allclose(diric_FS[: N_FS[0], : N_FS[1]], diric_FS_exact)
-       True
-
-    Notes
-    -----
-    Theory: :ref:`FFS_def`.
-
-    See Also
-    --------
-    :py:func:`~pyffs.util.ffsn_sample`, :py:func:`~pyffs.ffs.iffsn_comp`
-    """
-
-    axes, _ = _verify_ffsn_input(Phi, T, T_c, N_FS, axes)
-
-    # sequence of 1D FFS
-    Phi_FS = Phi.copy()
-    for d, ax in enumerate(axes):
-        Phi_FS = ffs(Phi_FS, T[d], T_c[d], N_FS[d], axis=ax)
-
-    return Phi_FS
-
-
-def iffsn_comp(Phi_FS, T, T_c, N_FS, axes=None):
-    r"""
-    Signal samples from Fourier Series coefficients of a D-dimension signal by performing D 1D iFFTs
-    along each dimension.
-
-    :py:func:`~pyffs.ffs.iffsn_comp` is basically the inverse of :py:func:`~pyffs.ffs.ffsn_comp` (
-    and :py:func:`~pyffs.ffs.ffsn`).
-
-    Parameters
-    ----------
-    Phi_FS : :py:class:`~numpy.ndarray`
-        (..., N_s1, N_s2, ..., N_sD, ...) FS coefficients in ascending order.
-    T : list(float)
-        Function period along each dimension.
-    T_c : list(float)
-        Function bandwidth along each dimension.
-    N_FS : list(int)
-        Period mid-point for each dimension.
-    axes : tuple
-        Dimensions of `Phi` along which function samples are stored.
-
-    Returns
-    -------
-    Phi : :py:class:`~numpy.ndarray`
-        (..., N_s1, N_s2, ..., N_sD, ...) matrices containing original function samples given to
-        :py:func:`~pyffs.ffs.ffsn`.
-
-        In short: :math:`(\text{iFFS} \circ \text{FFS})\{ x \} = x`.
-
-    Notes
-    -----
-    Theory: :ref:`FFS_def`.
-
-    See Also
-    --------
-    :py:func:`~pyffs.util.ffsn_sample`, :py:func:`~pyffs.ffs.ffsn_comp`
-    """
-
-    axes, _ = _verify_ffsn_input(Phi_FS, T, T_c, N_FS, axes)
-
-    # sequence of 1D iFFS
-    Phi = Phi_FS.copy()
-    for d, ax in enumerate(axes):
-        Phi = iffs(Phi, T[d], T_c[d], N_FS[d], axis=ax)
-
-    return Phi
-
-
 def ffsn(Phi, T, T_c, N_FS, axes=None):
     r"""
     Fourier Series coefficients from signal samples of a D-dimension signal by performing a
